WiggleLens/backend/requirements.txt
Flask==3.1.0
Pillow==9.5.0
moviepy==1.0.3
pytest==7.4.0
pytest-flask==1.2.0
flask-cors==4.0.0
